+++ packages/manager/src/features/Linodes/LinodeCreatev2/schemas.ts
@@ -1,0 +1,20 @@
+import { CreateLinodeSchema } from '@linode/validation';
+import { number, object } from 'yup';
+
+/**
+ * Extends the Linode Create schema to make `linode` required for the Clone Linode tab
+ */
+export const CreateLinodeByCloningSchema = CreateLinodeSchema.concat(
+  object({
+    linode: object().required('You must select a Linode to clone from'),
+  })
+);
+
+/**
+ * Extends the Linode Create schema to make backup_id required for the backups tab
+ */
+export const CreateLinodeFromBackupSchema = CreateLinodeSchema.concat(
+  object({
+    backup_id: number().required('You must select a Backup.'),
+  })
+);
